refactor: move viewcomponentatlassourcedatasetsselection to integrated object sd (#1004)
