+++ app/views/IntegratedObjectSourceDatasetsView/components/ViewComponentAtlasSourceDatasetsSelection/components/ComponentAtlasSourceDatasetsSelection/components/FormActions/formActions.tsx
@@ -1,7 +1,7 @@
 import { BUTTON_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/button";
 import { Button } from "@mui/material";
-import { FormActionsProps as CommonFormActionProps } from "../../../../../../../../../common/Form/components/FormActions/formActions";
-import { Actions } from "../../../../../../../../../common/Form/components/FormActions/formActions.styles";
+import { FormActionsProps as CommonFormActionProps } from "../../../../../../../../components/common/Form/components/FormActions/formActions";
+import { Actions } from "../../../../../../../../components/common/Form/components/FormActions/formActions.styles";
 
 interface FormActionsProps extends CommonFormActionProps {
   count: number;
