+++ app/views/IntegratedObjectSourceDatasetsView/components/ViewComponentAtlasSourceDatasetsSelection/components/ComponentAtlasSourceDatasetsSelection/components/FormContent/components/Table/common/constants.ts
@@ -7,7 +7,7 @@ import {
   SortingState,
   TableOptions,
 } from "@tanstack/react-table";
-import { HCAAtlasTrackerSourceDataset } from "../../../../../../../../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
+import { HCAAtlasTrackerSourceDataset } from "../../../../../../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
 
 const PUBLICATION_STRING = "publicationString";
 const TITLE = "title";
