+++ app/views/IntegratedObjectSourceDatasetsView/components/ViewComponentAtlasSourceDatasetsSelection/components/ComponentAtlasSourceDatasetsSelection/hooks/useComponentAtlasSourceDatasetsSelectionFormState.ts
@@ -1,10 +1,10 @@
 import { RowSelectionState } from "@tanstack/react-table";
 import { useEffect, useMemo } from "react";
-import { HCAAtlasTrackerSourceDataset } from "../../../../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
-import { FormMethod } from "../../../../../../../../../hooks/useForm/common/entities";
-import { PartialTableOptions } from "../../../../../../../../../hooks/useTableOptions";
-import { FIELD_NAME } from "../../../../../../ViewComponentAtlasSourceDatasetsSelection/common/constants";
-import { ComponentAtlasSourceDatasetsEditData } from "../../../../../../ViewComponentAtlasSourceDatasetsSelection/common/entities";
+import { HCAAtlasTrackerSourceDataset } from "../../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
+import { FormMethod } from "../../../../../../../hooks/useForm/common/entities";
+import { PartialTableOptions } from "../../../../../../../hooks/useTableOptions";
+import { FIELD_NAME } from "../../../common/constants";
+import { ComponentAtlasSourceDatasetsEditData } from "../../../common/entities";
 
 const SET_VALUE_OPTIONS = { shouldDirty: true, shouldValidate: true };
