+++ app/views/IntegratedObjectSourceDatasetsView/components/ViewComponentAtlasSourceDatasetsSelection/viewComponentAtlasSourceDatasetsSelection.styles.ts
@@ -1,6 +1,6 @@
 import { bpUpSm } from "@databiosphere/findable-ui/lib/styles/common/mixins/breakpoints";
 import styled from "@emotion/styled";
-import { StyledDialog as CommonDialog } from "../../../common/Dialog/dialog.styles";
+import { StyledDialog as CommonDialog } from "../../../../components/common/Dialog/dialog.styles";
 
 export const Dialog = styled(CommonDialog)`
   .MuiPaper-root {
